iIssue #40: Rearrange photos pages.
