+++ snekframe/analyse.py
@@ -4,7 +4,6 @@ import logging
 import os.path
 
 from .params import FILES_LOCATION, PHOTOS_LOCATION
-from .db import ExistingFiles, RUNTIME_ENGINE, RUNTIME_SESSION, PERSISTENT_SESSION, PhotoList, CurrentDisplay, NumPhotos, RuntimeBase
 
 import filetype.helpers
 import PIL.Image
